Revert "Updated the login route"
