--- routes/messages-api.js
@@ -1,4 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-module.exports = router;
