+++ routes/users.js
@@ -1,0 +1,85 @@
+/*
+ * All routes for Users are defined here
+ * Since this file is loaded in server.js into /users,
+ *   these routes are mounted onto /users
+ * See: https://expressjs.com/en/guide/using-middleware.html#middleware.router
+ */
+
+const express = require("express");
+const database = require("../db/database");
+
+const router = express.Router();
+
+// Create a new user
+router.post("/", (req, res) => {
+  const user = req.body;
+  database
+    .addUser(user)
+    .then((user) => {
+      if (!user) {
+        return res.send({ error: "error" });
+      }
+
+      req.session.userId = user.id;
+      res.send("🤗");
+    })
+    .catch((e) => res.send(e));
+});
+
+// Log a user in
+router.post("/login", (req, res) => {
+  const email = req.body.email;
+  const password = req.body.password;
+
+  database.getUserWithEmail(email).then((user) => {
+    if (!user) {
+      return res.send({ error: "no user with that id" });
+    }
+
+    if (password !== user.password) {
+      return res.send({ error: "error" });
+    }
+
+    req.session.userId = user.id;
+    res.send({
+      user: {
+        name: user.name,
+        email: user.email,
+        id: user.id,
+      },
+    });
+  });
+});
+
+// Log a user out
+router.post("/logout", (req, res) => {
+  req.session.userId = null;
+  res.send({});
+});
+
+// Return information about the current user (based on cookie value)
+router.get("/me", (req, res) => {
+  const userId = req.session.userId;
+  if (!userId) {
+    return res.send({ message: "not logged in" });
+  }
+
+  database
+    .getUserWithId(userId)
+    .then((user) => {
+      if (!user) {
+        return res.send({ error: "no user with that id" });
+      }
+
+      res.send({
+        user: {
+          name: user.name,
+          email: user.email,
+          id: userId,
+        },
+      });
+    })
+    .catch((e) => res.send(e));
+});
+
+module.exports = router;
