+++ server.js
@@ -28,28 +28,25 @@ app.use(express.static('public'));
 
 // Separated Routes for each Resource
 // Note: Feel free to replace the example routes below with your own
-const indexRoutes = require('./routes/indexRoutes');
-const usersRoutes = require('./routes/users-api');
-const itemsRoutes = require('./routes/items-api');
-const favoritesRoutes = require('./routes/favorites-api');
-const messagesRoutes = require('./routes/messages-api');
+const userApiRoutes = require('./routes/users-api');
+const widgetApiRoutes = require('./routes/widgets-api');
+const usersRoutes = require('./routes/users');
 
 // Mount all resource routes
 // Note: Feel free to replace the example routes below with your own
 // Note: Endpoints that return data (eg. JSON) usually start with `/api`
-app.use('/', indexRoutes);
-app.use('/api/users', usersRoutes);
-app.use('/api/items', itemsRoutes);
-app.use('/api/favorites', favoritesRoutes);
-app.use('/api/messages', messagesRoutes);
+app.use('/api/users', userApiRoutes);
+app.use('/api/widgets', widgetApiRoutes);
+app.use('/users', usersRoutes);
 // Note: mount other resources here, using the same pattern above
 
 // Home page
 // Warning: avoid creating more routes in this file!
 // Separate them into separate routes files (see above).
 
-
-
+app.get('/', (req, res) => {
+  res.render('index');
+});
 
 app.listen(PORT, () => {
   console.log(`Example app listening on port ${PORT}`);
