+++ views/index.ejs
@@ -1,122 +1,24 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
-  <head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
 
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
+<head>
+  <title>Home Page</title>
 
-    <title>index page</title>
-  </head>
+  <meta charset="UTF-8" >
+  <meta http-equiv="X-UA-Compatible" content="IE=edge" >
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" >
 
- 
-  <body>
+  <link rel="stylesheet" href="/vendor/normalize-4.1.1.css" />
+  <link rel="stylesheet" href="/vendor/border-box.css" />
+  <link rel="stylesheet" href="/styles/main.css" />
+  <link rel="stylesheet" href="/styles/layout.css" />
 
+  <script src="/vendor/jquery-3.0.0.js"></script>
+  <script defer src="/scripts/app.js"></script>
+</head>
 
-    <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-      <a class="navbar-brand" href="#">Buy/Sell</a>
-      <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-      </button>
-      <!-- <div class="collapse navbar-collapse" id="navbarNav"> -->
-        <div class="collapse navbar-collapse" id="navbarSupportedContent">
-        <ul class="navbar-nav mr-auto">
-          <li class="nav-item">
-            <a class="nav-link" href="/login">Login</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="/register">Register</a>
-          </li>
-        </ul>
-        <form class="form-inline my-2 my-lg-0">
-          <input class="form-control mr-sm-2" type="search" placeholder="Min Price" aria-label="Search">
-          <input class="form-control mr-sm-2" type="search" placeholder="Mix Price" aria-label="Search">
-          <button class="btn btn-outline-success my-2 my-sm-0" type="submit">Search</button>
-        </form>
-      </div>
-    </nav>
+<body>
+  <h1>Have fun! 🥔</h1>
+</body>
 
-   
-
-
-    <br><br>
-    <div class="card-deck">
-      <div class="card">
-        <img class="card-img-top" src="shoes.jpg" alt="shoes image">
-        <div class="card-body">
-          <h5 class="card-title">Ballet shoes</h5>
-          <p class="card-text"><small class="text-muted">Price : 230$</small></p>
-        </div>
-      </div>
-      <div class="card">
-        <img class="card-img-top" src="shoes.jpg" alt="shoes image">
-        <div class="card-body">
-          <h5 class="card-title">Ballet shoes</h5>
-          <p class="card-text"><small class="text-muted">Price : 230$</small></p>
-      </div>
-      </div>
-      <div class="card">
-        <img class="card-img-top" src="..." alt="shoes image">
-        <div class="card-body">
-          <h5 class="card-title">Ballet shoes</h5>
-          <p class="card-text"><small class="text-muted">Price : 230$</small></p>
-        </div>
-        
-      </div>
-      <div class="card">
-        <img class="card-img-top" src="..." alt="shoes image">
-        <div class="card-body">
-          <h5 class="card-title">Ballet shoes</h5>
-          <p class="card-text"><small class="text-muted">Price : 230$</small></p>
-        </div>
-        
-      </div>
-    </div>
-
-    <br><br>
-    <div class="card-deck">
-      <div class="card">
-        <img class="card-img-top" src="..." alt="shoes image">
-        <div class="card-body">
-          <h5 class="card-title">Ballet shoes</h5>
-          <p class="card-text"><small class="text-muted">Price : 230$</small></p>
-        </div>
-      </div>
-      <div class="card">
-        <img class="card-img-top" src="..." alt="shoes image">
-        <div class="card-body">
-          <h5 class="card-title">Ballet shoes</h5>
-          <p class="card-text"><small class="text-muted">Price : 230$</small></p>
-        </div>
-      </div>
-      <div class="card">
-        <img class="card-img-top" src="..." alt="shoes image">
-        <div class="card-body">
-          <h5 class="card-title">Ballet shoes</h5>
-          <p class="card-text"><small class="text-muted">Price : 230$</small></p>
-        </div>
-        
-      </div>
-      <div class="card">
-        <img class="card-img-top" src="..." alt="shoes image">
-        <div class="card-body">
-          <h5 class="card-title">Ballet shoes</h5>
-          <p class="card-text"><small class="text-muted">Price : 230$</small></p>
-        </div>
-        
-      </div>
-    </div>
-
-
-
-
-
-    <!-- Optional JavaScript -->
-    <!-- jQuery first, then Popper.js, then Bootstrap JS -->
-    <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.12.9/dist/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
-  </body>
 </html>
