--- views/messages.ejs
@@ -1,87 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <!-- Required meta tags -->
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-
-    <title>MY_ITEMS</title>
-  </head>
-
- 
-  <body>
-
-    <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-      <a class="navbar-brand" href="#">Buy/Sell</a>
-      <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-        <span class="navbar-toggler-icon"></span>
-      </button>
-      <!-- <div class="collapse navbar-collapse" id="navbarNav"> -->
-        <div class="collapse navbar-collapse" id="navbarSupportedContent">
-        <ul class="navbar-nav mr-auto">
-          <li class="nav-item">
-            <a class="nav-link" href="/add_items">ADD ITEMS</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="/my_items">MY ITEMS</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="/messages">MESSAGES</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="/favourites">FAVORITES</a>
-          </li>
-          <li class="nav-item">
-            <a class="nav-link" href="/index">LOGOUT</a>
-          </li>
-        </ul>
-        <form class="form-inline my-2 my-lg-0">
-          <input class="form-control mr-sm-2" type="search" placeholder="Min Price" aria-label="Search">
-          <input class="form-control mr-sm-2" type="search" placeholder="Mix Price" aria-label="Search">
-          <button class="btn btn-outline-success my-2 my-sm-0" type="submit">Search</button>
-        </form>
-      </div>
-    </nav>
-
-    <br><br>
-    <form>
-      <div class="form-group">
-        <label for="exampleFormControlTextarea1">messages</label>
-        <textarea class="form-control" id="exampleFormControlTextarea1" rows="3" cols="5"></textarea><br>
-        <button type="submit" class="btn btn-primary" cols ="3">Reply</button>
-        <button type="submit" class="btn btn-primary" cols = "3">Delete</button>
-      </div>
-      <div class="form-group">
-        <label for="exampleFormControlTextarea1">messages</label>
-        <textarea class="form-control" id="exampleFormControlTextarea1" rows="3" cols="5"></textarea><br>
-        <button type="submit" class="btn btn-primary" cols ="3">Reply</button>
-        <button type="submit" class="btn btn-primary" cols = "3">Delete</button>
-      </div>
-      <div class="form-group">
-        <label for="exampleFormControlTextarea1">messages</label>
-        <textarea class="form-control" id="exampleFormControlTextarea1" rows="3" cols="5"></textarea><br>
-        <button type="submit" class="btn btn-primary" cols ="3">Reply</button>
-        <button type="submit" class="btn btn-primary" cols = "3">Delete</button>
-      </div>
-      <div class="form-group">
-        <label for="exampleFormControlTextarea1">messages</label>
-        <textarea class="form-control" id="exampleFormControlTextarea1" rows="3" cols="5"></textarea><br>
-        <button type="submit" class="btn btn-primary" cols ="3">Reply</button>
-        <button type="submit" class="btn btn-primary" cols = "3">Delete</button>
-      </div>
-  </form>
-
-
-
-
-
-    <!-- Optional JavaScript -->
-    <!-- jQuery first, then Popper.js, then Bootstrap JS -->
-    <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.12.9/dist/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
-  </body>
-</html>
